# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GideonDB experimental API
  version: 0.0.1
  description: Phase 1 API; compatibility is not yet promised.
  contact:
    url: https://github.com/Bharanipbk/gideondb
servers:
  - url: http://127.0.0.1:6333/v1
security:
  - bearerAuth: []
paths:
...
  /collections:
    get:
      operationId: listCollections
      parameters:
        - {$ref: "#/components/parameters/PageLimit"}
        - {$ref: "#/components/parameters/PageCursor"}
      responses:
        "200":
          description: Collection configurations.
          content:
            application/json:
              schema:
                type: object
                properties:
                  collections:
                    type: array
                    items: {$ref: "#/components/schemas/CollectionConfig"}
                  next_cursor: {type: string}
                required: [collections, next_cursor]
    post:
      operationId: createCollection
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CollectionConfig"}
      responses:
        "201": {description: Collection created.}
        "400": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /collections/{name}:
    parameters:
      - {$ref: "#/components/parameters/CollectionName"}
    get:
      operationId: describeCollection
      responses:
        "200": {description: Collection description.}
        "404": {$ref: "#/components/responses/Error"}
    delete:
      operationId: deleteCollection
      responses:
        "204": {description: Collection deleted.}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
...
  /internal/replicas/{collection}/{shard}/append:
    parameters:
      - {name: collection, in: path, required: true, schema: {type: string}}
      - {name: shard, in: path, required: true, schema: {type: integer, minimum: 0}}
      - {name: X-GideonDB-Cluster-ID, in: header, required: true, schema: {type: string, pattern: "^[0-9a-f]{32}$"}}
      - {name: X-GideonDB-Target-Node-ID, in: header, required: true, schema: {type: string, pattern: "^[0-9a-f]{32}$"}}
      - {name: X-GideonDB-Metadata-Epoch, in: header, required: true, schema: {type: integer, minimum: 1}}
      - {name: X-GideonDB-Leader-Node-ID, in: header, required: true, schema: {type: string, pattern: "^[0-9a-f]{32}$"}}
      - {name: X-GideonDB-Leader-Term, in: header, required: true, schema: {type: integer, minimum: 1}}
    post:
      operationId: internalReplicaAppend
      description: Ordered, fenced follower WAL append; internal Phase 7 transport.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [sequence, replication_factor, records]
              properties:
                sequence: {type: integer, minimum: 1}
                replication_factor: {type: integer, minimum: 1}
                records: {type: array, minItems: 1, maxItems: 10000, items: {$ref: "#/components/schemas/Record"}}
      responses:
        "200": {description: Sequence durably appended or identical retry acknowledged.}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "410": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
  /internal/replicas/{collection}/{shard}/snapshot:
    parameters:
      - {name: collection, in: path, required: true, schema: {type: string}}
      - {name: shard, in: path, required: true, schema: {type: integer, minimum: 0}}
      - {name: X-GideonDB-Cluster-ID, in: header, required: true, schema: {type: string, pattern: "^[0-9a-f]{32}$"}}
      - {name: X-GideonDB-Target-Node-ID, in: header, required: true, schema: {type: string, pattern: "^[0-9a-f]{32}$"}}
      - {name: X-GideonDB-Metadata-Epoch, in: header, required: true, schema: {type: integer, minimum: 1}}
      - {name: X-GideonDB-Leader-Node-ID, in: header, required: true, schema: {type: string, pattern: "^[0-9a-f]{32}$"}}
      - {name: X-GideonDB-Leader-Term, in: header, required: true, schema: {type: integer, minimum: 1}}
    post:
      operationId: internalReplicaSnapshot
      description: Fenced complete-shard catch-up after a follower sequence gap or compaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [sequence, replication_factor, records]
              properties:
                sequence: {type: integer, minimum: 1}
                replication_factor: {type: integer, minimum: 1}
                records: {type: array, items: {$ref: "#/components/schemas/Record"}}
      responses:
        "200": {description: Snapshot durably installed at the supplied sequence.}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    CollectionName:
      name: name
      in: path
      required: true
      schema: {type: string, pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$"}
...
    PageLimit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 100}
    PageCursor:
      name: cursor
      in: query
      schema: {type: string}
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
...
    CollectionConfig:
      type: object
      additionalProperties: false
      properties:
        name: {type: string}
        dimension: {type: integer, minimum: 1}
        metric: {type: string, enum: [cosine, dot, l2]}
        shard_count: {type: integer, minimum: 1}
        index: {$ref: "#/components/schemas/IndexConfig"}
      required: [name, dimension, metric, shard_count]
    IndexConfig:
      type: object
      additionalProperties: false
      properties:
        type: {type: string, enum: [flat, hnsw], default: flat}
        m: {type: integer, minimum: 2, maximum: 128}
        ef_construction: {type: integer, minimum: 2, maximum: 10000}
        ef_search: {type: integer, minimum: 1, maximum: 10000}
    Record:
      type: object
      additionalProperties: false
      properties:
        id: {type: string}
        vector:
          type: array
          items: {type: number, format: float}
        sparse_vector: {type: object, minProperties: 1, maxProperties: 10000, additionalProperties: {type: number, format: float}}
        metadata: {type: object, additionalProperties: true}
        payload: {type: object, additionalProperties: true}
        namespace: {type: string}
        timestamp: {type: integer, format: int64, readOnly: true}
        version: {type: integer, format: uint64, readOnly: true}
      required: [id, vector]
...
    Error:
      type: object
      properties:
        code: {type: string}
        message: {type: string}
      required: [code, message]